requirements.txt
mysql-connector-python==8.3.0
sqlalchemy==2.0.27
